aq: Add AQ_AUTO_VARIANCE feature for performing adaptive quantization.

+++ source/encoder/ratecontrol.h
@@ -128,7 +128,7 @@ protected:
     double getQScale(RateControlEntry *rce, double rateFactor);
     double rateEstimateQscale(RateControlEntry *rce); // main logic for calculating QP based on ABR
     void accumPQpUpdate();
-    double acEnergyCu(TComPic* pic, uint32_t block_x, uint32_t block_y);
+    uint32_t acEnergyCu(TComPic* pic, uint32_t block_x, uint32_t block_y);
 
     void updateVbv(int64_t bits, RateControlEntry* rce);
     void updatePredictor(Predictor *p, double q, double var, double bits);
